requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
dnspython==2.2.1
idna==3.3
pymongo==4.1.0
python-dotenv==0.20.0
python-snappy==0.6.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
winkerberos==0.8.0
